compiles but segfaults when running

+++ functions.ml
@@ -32,10 +32,10 @@ let function_decls =
         (Bool, "visited", [Node, "n"]);    
         (Node, "updateData", [(Node, "n"); (String, "s")]); 
         (Node, "updateVisited", [(Node, "n"); (Bool, "b")]);
-        (Int, "cost", [Node, "n"]);
-        (Int, "incrementCost", [Node, "n"]);
-        (Int, "decrementCost", [Node, "n"]);   
-        (Int, "updateCost", [(Node, "n1"); (Int, "i")]); 
+        (Double, "cost", [Node, "n"]);
+        (Double, "incrementCost", [Node, "n"]);
+        (Double, "decrementCost", [Node, "n"]);   
+        (Double, "updateCost", [(Node, "n1"); (Double, "i")]); 
         (Node, "prec", [Node, "n"]);  
         (Node, "setPrec", [(Node, "n1"); (Node, "n2")]);    
         (Bool, "nodeEquals", [(Node, "n1"); (Node, "n2")]);
